Use double for quantity instead of int (#393)

+++ Libraries/SPTarkov.Server.Core/Helpers/Dialogue/Commando/SptCommands/TraderCommand/TraderSptCommand.cs
@@ -51,7 +51,7 @@ public class TraderSptCommand(
 
         var trader = result.Groups["trader"].Captures.Count > 0 ? result.Groups["trader"].Captures[0].Value : null;
         var command = result.Groups["command"].Captures.Count > 0 ? result.Groups["command"].Captures[0].Value : null;
-        var quantity = int.Parse(result.Groups["command"].Captures.Count > 0 ? result.Groups["quantity"].Captures[0].Value : "0");
+        var quantity = double.Parse(result.Groups["command"].Captures.Count > 0 ? result.Groups["quantity"].Captures[0].Value : "0");
 
         var dbTrader = _traderHelper.GetTraderByNickName(trader);
         if (dbTrader == null)
@@ -110,7 +110,7 @@ public class TraderSptCommand(
         return request.DialogId;
     }
 
-    protected ProfileChangeEvent CreateProfileChangeEvent(NotificationEventType profileChangeEventType, int quantity, string dbTraderId)
+    protected ProfileChangeEvent CreateProfileChangeEvent(NotificationEventType profileChangeEventType, double quantity, string dbTraderId)
     {
         return new ProfileChangeEvent
         {
